requirements.txt
flask
gunicorn
pillow
numpy
segmentation_models